* main: (37 commits)
  chore(build): replace Java device type stripper with Bash and awk (#37)
  Fix merge issues
  [SPHM-1549]: Sanitize JSON-API variable names in hmscript  and raised userlevel of saveObjectModel call to USER.
  [SPHM-1547][SPHM-1548][SPHM-1549]: Add some input-validation / -escaping
  hss_led: add missing PLATFORM_CCU3 define (#35)
  fix HmIP-RFUSB fwmap by defining firmware 4.4.18 the default (#34)
  downgrading HMServer.jar to working HomeMatic-only version (#33)
  add missing serial.tcl and tclload.tcl to add missing GetSerialNumber and tcl load statements. (#32)
  webui: cleanup obsolete files not required (#31)
  fix compiler warnings in hss_led, libeq3config and rfd (#30)
  libxmlparser: fix unsafe XMLNode array handling (#29)
  add CI workflow (#28)
  libXmlRpc: preserve dispatcher registration during reconnect failure (#27)
  improve version file generation in cmake build (#26)
  add fixes for devicetypes generation to abort build if necessary (#25)
  update tclrega and tclrpc to latest 1.4 versions (#24)
  bump rfd version to 1.2.0 and hs485d version to 2.6.0 (#23)
  update eq3_char_loop to version 1.5 including kernel v6 compatibility (#22)
  modify DEVDB create script to use lsort for deterministic creation (#21)
  add compat-libraries custom build target for libxmlparser/libXmlRpc builds only (#20)
  ...

# Conflicts:
#	src/hs485d/CMakeLists.txt
#	src/multimacd/CMakeLists.txt
#	src/rfd/CMakeLists.txt

+++ src/libeq3config/CCU2CommControllerMod/CCU2SerialFrameMod.cpp
@@ -157,7 +157,7 @@ void CCU2SerialFrameMod::deEscapeChar(char* c)
 bool CCU2SerialFrameMod::addFrameData(const std::string& frameData, std::string& leftOver)
 {
 	//LOG(Logger::LOG_ALL, "CCU2SerialFrame::addFrameData(): Processing %s", toDebugHexStr(frameData).c_str());
-		for(unsigned int i = 0; i < frameData.size() ; i++) {
+		for(size_t i = 0; i < frameData.size() ; i++) {
 			char c = frameData.at(i);
 			switch(c)
 			{
@@ -186,7 +186,7 @@ bool CCU2SerialFrameMod::addFrameData(const std::string& frameData, std::string&
 				}
 				if( ((int)payload.size() >= expectedMsgSize) && (expectedMsgSize != -1)) {
 					leftOver.clear();
-					if((int)frameData.size() > (i+1))  {
+					if(frameData.size() > (i+1))  {
 						leftOver.append( frameData.substr(i+1) );
 					}
 					payload = extractPayload(payload);
